feat: add semantic-release
ci: update release workflow for semantic-release

Refactor the release workflows for both alpha and prod branches to
utilize the cycjimmy/semantic-release-action for better release
management. This change simplifies the setup by removing the manual
installation of semantic-release and its plugins, and integrates
release note generation directly into the workflow.

Additionally, a new configuration file (.releaserc.js) is added to
define the release process and manage versioning based on branch
names. This enhances the automation of the release process and
ensures consistency across environments.